strip source chain back to make tests happy
