+++ hc_core/Cargo.toml
@@ -4,6 +4,9 @@ version = "0.1.0"
 authors = ["Nicolas Luck <nicolas@lucksus.org>"]
 
 [dependencies]
+serde = "1.0"
+serde_derive = "1.0"
+serde_json = "1.0"
 
 [features]
 # Treat warnings as a build error.
